+++ src/testing/test-helper.js
@@ -409,7 +409,7 @@ function verifyAccessAssignments(expectedAccessAssignments) {
   }
 }
 
-function verifyOperationChannelsAssigned(doc, oldDoc, expectedChannels) {
+function verifyOperationChannelsAssigned(doc, expectedChannels) {
   if (exports.channel.callCount !== 1) {
     assert.fail('Document channels were not assigned');
   }
@@ -464,7 +464,7 @@ function verifyAuthorization(expectedAuthorization) {
 }
 
 function verifyDocumentAccepted(doc, oldDoc, expectedAuthorization, expectedAccessAssignments) {
-  exports.syncFunction(doc, oldDoc);
+  exports.syncFunction(doc, oldDoc || null);
 
   if (expectedAccessAssignments) {
     verifyAccessAssignments(expectedAccessAssignments);
@@ -472,11 +472,11 @@ function verifyDocumentAccepted(doc, oldDoc, expectedAuthorization, expectedAcce
 
   const expectedOperationChannels = verifyAuthorization(expectedAuthorization);
 
-  verifyOperationChannelsAssigned(doc, oldDoc, expectedOperationChannels);
+  verifyOperationChannelsAssigned(doc, expectedOperationChannels);
 }
 
 function verifyDocumentCreated(doc, expectedAuthorization, expectedAccessAssignments) {
-  verifyDocumentAccepted(doc, void 0, expectedAuthorization || defaultWriteChannel, expectedAccessAssignments);
+  verifyDocumentAccepted(doc, null, expectedAuthorization || defaultWriteChannel, expectedAccessAssignments);
 }
 
 function verifyDocumentReplaced(doc, oldDoc, expectedAuthorization, expectedAccessAssignments) {
@@ -490,7 +490,7 @@ function verifyDocumentDeleted(oldDoc, expectedAuthorization) {
 function verifyDocumentRejected(doc, oldDoc, docType, expectedErrorMessages, expectedAuthorization) {
   let syncFuncError = null;
   try {
-    exports.syncFunction(doc, oldDoc);
+    exports.syncFunction(doc, oldDoc || null);
   } catch (ex) {
     syncFuncError = ex;
   }
@@ -506,7 +506,7 @@ function verifyDocumentRejected(doc, oldDoc, docType, expectedErrorMessages, exp
 }
 
 function verifyDocumentNotCreated(doc, docType, expectedErrorMessages, expectedAuthorization) {
-  verifyDocumentRejected(doc, void 0, docType, expectedErrorMessages, expectedAuthorization || defaultWriteChannel);
+  verifyDocumentRejected(doc, null, docType, expectedErrorMessages, expectedAuthorization || defaultWriteChannel);
 }
 
 function verifyDocumentNotReplaced(doc, oldDoc, docType, expectedErrorMessages, expectedAuthorization) {
@@ -583,7 +583,7 @@ function verifyAccessDenied(doc, oldDoc, expectedAuthorization) {
 
   let syncFuncError = null;
   try {
-    exports.syncFunction(doc, oldDoc);
+    exports.syncFunction(doc, oldDoc || null);
   } catch (ex) {
     syncFuncError = ex;
   }
@@ -627,7 +627,7 @@ function verifyAccessDenied(doc, oldDoc, expectedAuthorization) {
 function verifyUnknownDocumentType(doc, oldDoc) {
   let syncFuncError = null;
   try {
-    exports.syncFunction(doc, oldDoc);
+    exports.syncFunction(doc, oldDoc || null);
   } catch (ex) {
     syncFuncError = ex;
   }